requirements.txt
idleiss >= 0.0.1
discord.py >= 1.7.3